programs, performance: fix clang11 compilation bug in timer_test_periodic
Fix the clang11 compilation error

timer_test_periodic.c:1652:54: error: comparison of integers of different signs:
'long long' and 'uint64_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
        if (g_options.first_ns &&
            llabs(g_options.first_ns) < g_options.min_period_ns)
            ~~~~~~~~~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~

by casting the llabs() return value to uint64_t before the comparison:
        if (g_options.first_ns &&
            (uint64_t)llabs(g_options.first_ns) < g_options.min_period_ns)

+++ programs/performance/timer_test_periodic.c
@@ -1649,7 +1649,7 @@ void test_start(appl_conf_t *const appl_conf)
 		if (g_options.min_period_ns < res_capa.min_tmo)
 			g_options.min_period_ns = res_capa.min_tmo;
 	}
-	if (g_options.first_ns && llabs(g_options.first_ns) < g_options.min_period_ns) {
+	if (g_options.first_ns && (uint64_t)llabs(g_options.first_ns) < g_options.min_period_ns) {
 		if (g_options.first_ns < 0)
 			g_options.first_ns = 0 - g_options.min_period_ns;
 		else
